make C & C++ not duplicate typedef struct foo foo
Previously, if you had a struct definition with a typedef of the form

typedef struct foo { ... } foo;

you would get two definitions listed for foo in the tags file: one on
the line of the first 'foo', one on the second. Now you only get the
first one. This change only affects things where the struct definition
is in the typedef line (and unions, and enums). Definitions of the form

typedef struct foo foo;
typedef struct foo bar;
typedef struct foo { ... } bar;

are all unaffected, and continue to result in two definitions in the
tags file.